Add files via upload
Version v0.02 changed the VBO and EBO to take vectors and not static arrays, makes it much more efficient to draw multiple entities. Each triangle I draw on the screen has its normalized coordinates stored in the vertexdata.txt file, which updates as you add more and more rectangles, and resets everytime the app is closed. 

To-do for v0.03: 
- automate the draw index process
- place rectangles on the screen with the mouse